Convert part 2 to use interfaces and bitflags
