+++ curriculum/part-2/1.go
@@ -1,28 +1,18 @@
-package entity
+package interfaces
 
 import (
 	"github.com/kettek/goro"
 )
 
-// Entity is a type that represents an active entity in the world.
-type Entity struct {
-	X, Y  int
-	Rune  rune
-	Style goro.Style
-}
-
-// Move moves the entity by a given amount.
-func (e *Entity) Move(x, y int) {
-	e.X += x
-	e.Y += y
-}
-
-// NewEntity returns a pointer to a newly created Entity.
-func NewEntity(x int, y int, r rune, s goro.Style) *Entity {
-	return &Entity{
-		X:     x,
-		Y:     y,
-		Rune:  r,
-		Style: s,
-	}
+type Entity interface {
+	// Basic property accessors
+	X() int
+	SetX(int)
+	Y() int
+	SetY(int)
+	Rune(rune)
+	SetRune(rune)
+	Style() goro.Style
+	SetStyle(goro.Style)
+	// More complex methods
 }
