+++ curriculum/part-2/1.md
@@ -1,56 +1,42 @@
-# The Generic Entity
-Now we can move on to creating a generic Entity type that we will use for storing data such as position, rune, and style.
+# The Entity Interface
+We begin defining our Entity type by first creating an interface. Create a new directory in your project root called **interfaces**. Within this directory, create a new file called **entity.go**.
 
-However, to begin with, let us create a new directory in our project directory called **entity** that will store our Entity type and any future entity-related code. Now create a new file in the **entity** directory called **entity.go** and fill it with the code to the right.
+The interface we create will be the data type through which we access our concrete Entity types underneath. This gives us the ability to extend the functionality of an Entity beyond a single implementation of it.
 
 ## Code Analysis
-
 ```
-package entity
++package interfaces
+```
+We define our new **Entity** interface to be part of the interfaces package.
 
+---
+```
 import (
-  "github.com/kettek/goro"
+	"github.com/kettek/goro"
 )
-```
-Here we are declaring our package to be **entity** rather than main since we will be importing this package for use in our main program. Additionally, we are importing goro for use, as we'll be using `goro.Style` within this file.
+``
+We must also import the goro library so we have access to **goro.Style**.
 
 ---
 ```
-// Entity is a type that represents an active entity in the world.
-type Entity struct {
-  X, Y int
-  Rune rune
-  Style goro.Style
+type Entity interface {
+	// Basic property accessors
+	X() int
+	SetX(int)
+	Y() int
+	SetY(int)
+	Rune(rune)
+	SetRune(rune)
+	Style() goro.Style
+	SetStyle(goro.Style)
+	// More complex methods
+  Move(int, int)
 }
 ```
-Now we have our Entity type that contains the pertinent data for moving and rendering it.
-
----
-```
-// Move moves the entity by a given amount.
-func (e *Entity) Move(x, y int) {
-  e.X += x
-  e.Y += y
-}
-```
-This is a move method for an Entity that will change its X and Y values by adding the provided x and y parameters to them.
-
----
-```
-// NewEntity returns a pointer to a newly created Entity.
-func NewEntity(x int, y int, r rune, s goro.Style) *Entity {
-  return &Entity{
-    X:     x,
-    Y:     y,
-    Rune:  r,
-    Style: s,
-  }
-}
-```
-`NewEntity(...)` is simply a helper function for constructing a new entity. Instead of this, you could use write the `&Entity{...}` code itself wherever NewEntity would be used.
+Here we have our Entity interface definition. We are defining setters and getters(accessors) for the Entity's x, y, rune, and style properties. We are also defining the `Move(int, int)` method that we will be using to adjust our Entity's position by a given amount.
 
 ## Topics Covered
-Here we have simply created the Entity type and defined a single method and a utility function for creating it.
+Implementing a basic Entity interface that we will later use.
 
 ## What's Next
-With this, we can now actually use our Entity type in our main program file.
+Now we can move on to defining a concrete Entity type that implements our Entity interface.
