+++ curriculum/part-2/2.md
@@ -1,50 +1,110 @@
-# Using the Entity
-We can now return to our **main.go** file and use the `Entity` type.
+# The Entity Type
+Now we can move on to creating a generic Entity type that we will use for storing data such as position, rune, and style.
+
+However, to begin with, let us create a new directory in our project directory called **entity** that will store our Entity type and any future entity-related code. Now create a new file in the **entity** directory called **entity.go** and fill it with the code to the right.
 
 ## Code Analysis
-As per usual, we will only be focusing on changes or removals to our base code.
 
 ```
-import (
-	"github.com/kettek/goro"
-	"log"
++package entity
 +
-+	"myproject/entity"
-)
++import (
++ "github.com/kettek/goro"
++
++	"myproject/interfaces"
++)
 ```
-Here we have added an import line for using our new **entity** package. Be sure to replace **myproject** with whatever you named your module in [../part-1/1](Part 1:1).
+Here we are declaring our package to be **entity** rather than main since we will be importing this package for use in our main program. Additionally, we are importing goro for use, as we'll be using `goro.Style` within this file. We are also importing our **interfaces** package so we can properly use the Entity interface. Be sure to replace **myproject** with whatever you named your module in [../part-1/1](Part 1:1).
 
 ---
 ```
-		// Our initial variables.
--		playerX, playerY := screen.Columns/2, screen.Rows/2
-+		player := entity.NewEntity(screen.Columns/2, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorWhite})
++// Entity is a type that represents an active entity in the world.
++type Entity struct {
++  x, y int
++  rune rune
++  style goro.Style
++}
 ```
-
-Here we have replaced our `playerX` and `playerY` variables with a single `player` variable that is an instance of an `Entity` returned from our `entity.NewEntity(...)` helper function. We have additionally introduced an explicit color to use for drawing the player. If you wish to experiment, there are other default colors available to use that are listed [here](https://godoc.org/github.com/kettek/goro#pkg-variables).
-
----
-```
-			// Draw screen.
--			screen.DrawRune(playerX, playerY, '@', goro.Style{})
-+			screen.DrawRune(player.X, player.Y, player.Rune, player.Style)
-			screen.Flush()
--			screen.DrawRune(playerX, playerY, ' ', goro.Style{})
-+			screen.DrawRune(player.X, player.Y, ' ', goro.Style{})
-```
-This is a fairly straight-forward change to use the `player` variable instead of the `playerX` and `playerY` variables. We have also changed our `screen.DrawRune(...)` call to use the `Rune` and `Style` properties of our `player` variable.
+Now we have our Entity type that contains the pertinent data for moving and rendering it.
 
 ---
 ```
-				case ActionMove:
--					playerX += action.X
--					playerY += action.Y
-+					player.Move(action.X, action.Y)
++// NewEntity returns an interfaces.Entity to our underlying Entity type.
++func NewEntity(x int, y int, r rune, s goro.Style) interfaces.Entity {
++  return &Entity{
++    x:     x,
++    y:     y,
++    rune:  r,
++    style: s,
++  }
++}
 ```
-Of course we must also change our `ActionMove` handling so as to use the new entity system. As can be seen, we are simply replacing our calls to adjust the `playerX` and `playerY` variables with a call to our `player.Move(...)` method.
+`NewEntity(...)` is simply a helper function for constructing a new entity pointer and returning it as an interface.Entity. Instead of this, you could use write `interfaces.Entity{&Entity{...}}` wherever NewEntity would be used.
+
+---
+```
++// X returns the entity's X position.
++func (e *Entity) X() int {
++	return e.x
++}
++
++// SetX sets the entity's X position.
++func (e *Entity) SetX(x int) {
++	e.x = x
++}
++
++// Y returns the entity's Y position.
++func (e *Entity) Y() int {
++	return e.y
++}
++
++// SetY sets the entity's Y position.
++func (e *Entity) SetX(y int) {
++	e.y = y
++}
+```
+Here we have the getters and setters for our x and y entity properties. These are simply methods through which we can modify or access our Entity's underlying data.
+
+---
+```
++// Rune returns the entity's rune.
++func (e *Entity) Rune() rune {
++	return e.rune
++}
++
++// SetRune sets the entity's rune.
++func (e *Entity) SetRune(r rune) {
++	e.rune = r
++}
+```
+Here we have the getter and setter for our Entity's rune.
+
+---
+```
++// Style returns the entity's goro.Style.
++func (e *Entity) Style() goro.Style {
++	return e.style
++}
++
++// SetStyle sets the entity's goro.Style.
++func (e *Entity) SetStyle(s goro.Style) {
++	e.style = s
++}
+```
+And finally we have our Style getter and setter.
+
+---
+```
++// Move moves the entity by a given amount.
++func (e *Entity) Move(x, y int) {
++  e.x += x
++  e.y += y
++}
+```
+This is a move method for an Entity that will change its X and Y values by adding the provided x and y parameters to them.
 
 ## Topics Covered
-We have covered converting our old player variables with a single player variable that is of the `Entity` type. This abstraction will allow us to also add NPCs that can move around as well (eventually).
+Here we have created our Entity type that fully implements our Entity interface.
 
 ## What's Next
-Now we must implement some new ways of doing rendering so that we will be able to render multiple entities.
+With this, we can now actually use our Entity interface and type in our main program file.
