+++ curriculum/part-2/3.go
@@ -1,32 +1,44 @@
 package main
 
 import (
+	"log"
+
 	"github.com/kettek/goro"
 
 	"myproject/entity"
 )
 
-// DrawAll draws all entities to the screen and flushes it.
-func DrawAll(screen *goro.Screen, entities []*entity.Entity) {
-	for _, entity := range entities {
-		DrawEntity(screen, entity)
+func main() {
+	// Initialize goro!
+	if err := goro.InitTCell(); err != nil {
+		log.Fatal(err)
 	}
-	screen.Flush()
-}
 
-// ClearAll clears all entities from the screen.
-func ClearAll(screen *goro.Screen, entities []*entity.Entity) {
-	for _, entity := range entities {
-		ClearEntity(screen, entity)
-	}
-}
+	goro.Run(func(screen *goro.Screen) {
+		// Screen configuration.
+		screen.SetTitle("My Roguelike")
 
-// DrawEntity draws a given entity to the screen.
-func DrawEntity(screen *goro.Screen, e *entity.Entity) {
-	screen.DrawRune(e.X, e.Y, e.Rune, e.Style)
-}
+		// Our initial variables.
+		player := entity.NewEntity(screen.Columns/2, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorWhite})
 
-// ClearEntity clears a given entity from the screen.
-func ClearEntity(screen *goro.Screen, e *entity.Entity) {
-	screen.DrawRune(e.X, e.Y, ' ', goro.Style{})
+		for {
+			// Draw screen.
+			screen.DrawRune(player.X(), player.Y(), player.Rune(), player.Style())
+			screen.Flush()
+			screen.DrawRune(player.X(), player.Y(), ' ', goro.Style{})
+
+			// Handle events.
+			switch event := screen.WaitEvent().(type) {
+			case goro.EventKey:
+				switch action := handleKeyEvent(event).(type) {
+				case ActionMove:
+					player.Move(action.X, action.Y)
+				case ActionQuit:
+					goro.Quit()
+				}
+			case goro.EventQuit:
+				return
+			}
+		}
+	})
 }
