+++ curriculum/part-2/3.md
@@ -1,66 +1,50 @@
-# Implementing Rendering
-We must now create a new file called **render.go**. This will hold functions for drawing and clearing entities and other types from the screen. Put the code to the right in that file.
+# Using the Entity
+We can now return to our **main.go** file and use the `Entity` type and interface.
 
 ## Code Analysis
+As per usual, we will only be focusing on changes or removals to our base code.
 
 ```
-package main
-
 import (
 	"github.com/kettek/goro"
-
-	"myproject/entity"
+	"log"
++
++	"myproject/entity"
 )
 ```
-Here we have our standard package and imports. We are also importing our own entity package for use within rendering.
+Here we have added an import line for using our new **entity** package. Be sure to replace **myproject** with whatever you named your module in [../part-1/1](Part 1:1).
 
 ---
 ```
-// DrawAll draws all entities to the screen and flushes it.
-func DrawAll(screen *goro.Screen, entities []*entity.Entity) {
-	for _, entity := range entities {
-		DrawEntity(screen, entity)
-	}
-	screen.Flush()
-}
+		// Our initial variables.
+-		playerX, playerY := screen.Columns/2, screen.Rows/2
++		player := entity.NewEntity(screen.Columns/2, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorWhite})
 ```
-As described by the function comment, the code here draws all the entities to the screen. The first argument to `DrawAll(...)` is a pointer to our screen and the second argument is a slice of Entity pointers. This slice will contain pointers to our player and any NPCs we create.
 
-Within the `for range` loop, we are making calls to a separate `DrawEntity(...)` function. We will be defining this shortly.
-
-Of course we also end with a call to `screen.Flush()` so as to ensure everything is displayed properly.
+Here we have replaced our `playerX` and `playerY` variables with a single `player` variable that is an instance of an `interfaces.Entity` returned from our `entity.NewEntity(...)` helper function. We have additionally introduced an explicit color to use for drawing the player. If you wish to experiment, there are other default colors available to use that are listed [here](https://godoc.org/github.com/kettek/goro#pkg-variables).
 
 ---
 ```
-// ClearAll clears all entities from the screen.
-func ClearAll(screen *goro.Screen, entities []*entity.Entity) {
-	for _, entity := range entities {
-		ClearEntity(screen, entity)
-	}
-}
+			// Draw screen.
+-			screen.DrawRune(playerX, playerY, '@', goro.Style{})
++			screen.DrawRune(player.X(), player.Y(), player.Rune(), player.Style())
+			screen.Flush()
+-			screen.DrawRune(playerX, playerY, ' ', goro.Style{})
++			screen.DrawRune(player.X(), player.Y(), ' ', goro.Style{})
 ```
-`ClearAll(...)` can be considered as the separate half to the `DrawAll(...)` function as it will clear the locations of the passed entities but will not call `Screen.Flush()`. If you recall how our current rendering works, it is **draw -> flush -> clear**. We are setting up to do the same thing here but with multiple entities.
+This is a fairly straight-forward change to use the `player`'s `X()` and `Y()` getters instead of the `playerX` and `playerY` variables. We have also changed our `screen.DrawRune(...)` call to use the `Rune()` and `Style()` getters of our `player` variable.
 
 ---
 ```
-// DrawEntity draws a given entity to the screen.
-func DrawEntity(screen *goro.Screen, e *entity.Entity) {
-	screen.DrawRune(e.X, e.Y, e.Rune, e.Style)
-}
+				case ActionMove:
+-					playerX += action.X
+-					playerY += action.Y
++					player.Move(action.X, action.Y)
 ```
-Here we have our `DrawEntity(...)` function that draws a given entity to the screen. It should look familiar as the contents of this function is how we are drawing the player to the screen in the previous lesson.
-
----
-```
-// ClearEntity clears a given entity from the screen.
-func ClearEntity(screen *goro.Screen, e *entity.Entity) {
-	screen.DrawRune(e.X, e.Y, ' ', goro.Style{})
-}
-```
-And here we are clearing the rune at an entity's position. This should also look familiar for the same reason as above.
+Of course we must also change our `ActionMove` handling so as to use the new Entity interface. As can be seen, we are simply replacing our calls to adjust the `playerX` and `playerY` variables with a call to our `player.Move(...)` method.
 
 ## Topics Covered
-We have created functions for rendering multiple entities to the screen.
+We have covered converting our old player variables with a single player variable that is of the `interfaces.Entity` type. This abstraction will allow us to also add NPCs that can move around as well (eventually).
 
 ## What's Next
-Now we can return to our main program and update it so that we have more than just the player rendered to the screen.
+Now we must implement some new ways of doing rendering so that we will be able to render multiple entities.
