+++ curriculum/part-2/4.md
@@ -1,38 +1,66 @@
-# Multiple Entities and Rendering
-Now we can return to our **main.go** file and adjust it to use our new rendering functions.
+# Implementing Rendering
+We must now create a new file called **render.go**. This will hold functions for drawing and clearing entities and other types from the screen. Put the code to the right in that file.
 
 ## Code Analysis
 
 ```
-		// Our initial variables.
-		player := entity.NewEntity(screen.Columns/2, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorWhite})
-+		npc := entity.NewEntity(screen.Columns/2-5, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorYellow})
-+
-+		entities := []*entity.Entity{
-+			player,
-+			npc,
-+		}
-```
-Here we begin by creating a new variable called `npc` that is almost the same as our `player` entity but differs in position and foreground color.
+package main
 
-Next, we create the variable, `entities` that refers to a slice of Entity pointers. We construct this slice in-place by populating it with our `player` and `npc` variables.
+import (
+	"github.com/kettek/goro"
+
+	"myproject/interfaces"
+)
+```
+Here we have our standard package and imports. We are also importing our own interfaces package so we can use our Entity interface during rendering.
 
 ---
 ```
-		for {
-			// Draw screen.
--			screen.DrawRune(player.X, player.Y, player.Rune, player.Style)
--			screen.Flush()
--			screen.DrawRune(player.X, player.Y, ' ', goro.Style{})
-+			DrawAll(screen, entities)
-+			ClearAll(screen, entities)
+// DrawAll draws all entities to the screen and flushes it.
+func DrawAll(screen *goro.Screen, entities []interfaces.Entity) {
+	for _, entity := range entities {
+		DrawEntity(screen, entity)
+	}
+	screen.Flush()
+}
 ```
-Now we remove our old calls to render the player to the screen and replace it with our newly created `DrawAll(...)` and `ClearAll(...)` functions.
+As described by the function comment, the code here draws all the entities to the screen. The first argument to `DrawAll(...)` is a pointer to our screen and the second argument is a slice of Entity interfaces. This slice will contain the Entity interfaces to our player and any NPCs we create.
 
-At this point, if you rebuild the program and run it, you should see two @ symbols on the screen and be able to move around the one representing the player.
+Within the `for range` loop, we are making calls to a separate `DrawEntity(...)` function. We will be defining this shortly.
+
+Of course we also end with a call to `screen.Flush()` so as to ensure everything is displayed properly.
+
+---
+```
+// ClearAll clears all entities from the screen.
+func ClearAll(screen *goro.Screen, entities []interfaces.Entity) {
+	for _, entity := range entities {
+		ClearEntity(screen, entity)
+	}
+}
+```
+`ClearAll(...)` can be considered as the separate half to the `DrawAll(...)` function as it will clear the locations of the passed entities but will not call `Screen.Flush()`. If you recall how our current rendering works, it is **draw -> flush -> clear**. We are setting up to do the same thing here but with multiple entities.
+
+---
+```
+// DrawEntity draws a given entity to the screen.
+func DrawEntity(screen *goro.Screen, e interfaces.Entity) {
+	screen.DrawRune(e.X(), e.Y(), e.Rune(), e.Style())
+}
+```
+Here we have our `DrawEntity(...)` function that draws a given entity to the screen. It should look familiar as the contents of this function is how we are drawing the player to the screen in the previous lesson.
+
+---
+```
+// ClearEntity clears a given entity from the screen.
+func ClearEntity(screen *goro.Screen, e interfaces.Entity) {
+	screen.DrawRune(e.X(), e.Y(), ' ', goro.Style{})
+}
+```
+And here we are clearing the rune at an entity's position. This should also look familiar for the same reason as above.
 
 ## Topics Covered
-We have created a list of entities that we can pass around and are now using properly abstracted rendering calls.
+We have created functions for rendering multiple entities to the screen.
 
 ## What's Next
-Now we will shift gears a bit and get into creating the Tile type followed by a GameMap type.
+Now we can return to our main program and update it so that we have more than just the player rendered to the screen.
