+++ curriculum/part-2/5.go
@@ -1,7 +1,50 @@
-package mapping
+package main
 
-// Tile represents the state of a given location in a GameMap.
-type Tile struct {
-	BlockMovement bool
-	BlockSight    bool
+import (
+	"log"
+
+	"github.com/kettek/goro"
+
+	"myproject/entity"
+	"myproject/interfaces"
+)
+
+func main() {
+	// Initialize goro!
+	if err := goro.InitTCell(); err != nil {
+		log.Fatal(err)
+	}
+
+	goro.Run(func(screen *goro.Screen) {
+		// Screen configuration.
+		screen.SetTitle("My Roguelike")
+
+		// Our initial variables.
+		player := entity.NewEntity(screen.Columns/2, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorWhite})
+		npc := entity.NewEntity(screen.Columns/2-5, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorYellow})
+
+		entities := []interfaces.Entity{
+			player,
+			npc,
+		}
+
+		for {
+			// Draw screen.
+			DrawAll(screen, entities)
+			ClearAll(screen, entities)
+
+			// Handle events.
+			switch event := screen.WaitEvent().(type) {
+			case goro.EventKey:
+				switch action := handleKeyEvent(event).(type) {
+				case ActionMove:
+					player.Move(action.X, action.Y)
+				case ActionQuit:
+					goro.Quit()
+				}
+			case goro.EventQuit:
+				return
+			}
+		}
+	})
 }
