+++ curriculum/part-2/5.md
@@ -1,26 +1,50 @@
-# Mapping: The Tile
-So as to ensure a degree of organization is maintained, we must first create a new directory within your project directory titled **mapping**. Within this directory, create a file titled **tile.go** and place the code to the right within it.
+# Multiple Entities and Rendering
+Now we can return to our **main.go** file and adjust it to use our new rendering functions.
 
 ## Code Analysis
 
 ```
-package mapping
-```
-Here we are using a different package name than "main". Just as we refer to any goro library by prepending "goro." to it, when referring to a `Tile` we will prepend "mapping.".
+import (
+	"log"
 
+	"github.com/kettek/goro"
+
+	"myproject/entity"
++	"myproject/interfaces"
+)
+```
+Here we have added our **interfaces** package to our imports so we can use it later.
+
+```
+		// Our initial variables.
+		player := entity.NewEntity(screen.Columns/2, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorWhite})
++		npc := entity.NewEntity(screen.Columns/2-5, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorYellow})
++
++		entities := []interfaces.Entity{
++			player,
++			npc,
++		}
+```
+Here we begin by creating a new variable called `npc` that is almost the same as our `player` entity but differs in position and foreground color.
+
+Next, we create the variable, `entities` that refers to a slice of Entity interfaces. We construct this slice in-place by populating it with our `player` and `npc` variables.
 
 ---
 ```
-// Tile represents the state of a given location in a GameMap.
-type Tile struct {
-	BlockMovement bool
-	BlockSight    bool
-}
+		for {
+			// Draw screen.
+-			screen.DrawRune(player.X(), player.Y(), player.Rune(), player.Style())
+-			screen.Flush()
+-			screen.DrawRune(player.X(), player.Y(), ' ', goro.Style{})
++			DrawAll(screen, entities)
++			ClearAll(screen, entities)
 ```
-Here is our Tile type. We have two simple boolean values as properties here: **BlockMovement** for blocking entity movement and **BlockSight** for blocking field-of-view (once we get there).
+Now we remove our old calls to render the player to the screen and replace it with our newly created `DrawAll(...)` and `ClearAll(...)` functions.
+
+At this point, if you rebuild the program and run it, you should see two @ symbols on the screen and be able to move around the one representing the player.
 
 ## Topics Covered
-We have created a new package for holding types and functions corresponding to mapping and we have also defined a very basic Tile type.
+We have created a list of entities that we can pass around and are now using properly abstracted rendering calls.
 
 ## What's Next
-Now we can move to creating a `GameMap` type using our Tile type.
+Now we will shift gears a bit and get into creating the Tile type followed by a GameMap type.
