+++ curriculum/part-2/6.md
@@ -1,68 +1,34 @@
-# Mapping: The GameMap
-With our Tile type now in place, we can create a new file in our **mapping** directory called **gamemap.go**. This GameMap will hold the map structure and important functions for map generation.
+# Mapping: The Tile
+So as to ensure a degree of organization is maintained, we must first create a new directory within your project directory titled **mapping**. Within this directory, create a file titled **tile.go** and place the code to the right within it.
 
 ## Code Analysis
 
 ```
-package mapping
++package mapping
 ```
-Here we are declaring our GameMap type to exist in our mapping package.
+Here we are using a different package name than "main". Just as we refer to any goro library by prepending "goro." to it, when referring to a `Tile` we will prepend "mapping.".
 
 ---
 ```
-// GameMap is our map data type.
-type GameMap struct {
-	Width, Height int
-	Tiles         [][]Tile
-}
++// Tile represents the state of a given location in a GameMap.
++type Tile struct {
++	Flags uint
++}
 ```
-This is our `GameMap` type that contains the map's width, height, and two-dimensional slice of Tile types.
+Here is our Tile type. We have a single unsigned integer value as a property here: **Flags**. We will use this single integer to store multiple boolean values as bitflags. How this works is that the underlying binary data of an integer is composed of multiple bits that can either be on(1) or off(0). By turning particular bits on or off, we can represent multiple boolean values within a single integer. Although we could use regular booleans, we gain the ability to do multiple-matching and filtering through the use of [bitwise operators](https://yourbasic.org/golang/bitwise-operator-cheat-sheet/).
 
 ---
 ```
-// Initialize initializes a GameMap's Tiles to match its Width and Height. It also sets up some coordinates to block movement and sight.
-func (g *GameMap) Initialize() {
-	g.Tiles = make([][]Tile, g.Width)
-
-	for x := range g.Tiles {
-		g.Tiles[x] = make([]Tile, g.Height)
-	}
++// Our Tile's flags.
++const (
++	BlockMovement = 1 << iota
++	BlockSight
++)
 ```
-Here is our GameMap's `Initialize()` function that we use to create its two-dimensional slice of Tile types. We are first using golang's built-in [https://tour.golang.org/moretypes/13](make(...)) function to create a slice of slices of Tiles up to the map's width. After this we iterate over each column of `g.Tiles` and create a slice of Tiles within each up to the map's height.
-
----
-```
-	g.Tiles[30][22] = Tile{
-		BlockMovement: true,
-		BlockSight:    true,
-	}
-	g.Tiles[31][22] = Tile{
-		BlockMovement: true,
-		BlockSight:    true,
-	}
-	g.Tiles[32][22] = Tile{
-		BlockMovement: true,
-		BlockSight:    true,
-	}
-}
-```
-Now we are setting three tiles at **30x22**, **31x22**, and **32x22** to both block movement and sight. And, finally, we are closing the function with the curly brace on the last line.
-
----
-```
-// IsBlocked returns if the given coordinates are blocking movement.
-func (g *GameMap) IsBlocked(x, y int) bool {
-  // Always block if outside our GameMap's bounds.
-  if x < 0 || x >= g.Width || y < 0 || y >= g.Height {
-    return true
-  }
-	return g.Tiles[x][y].BlockMovement
-}
-```
-`IsBlocked(...)` simply returns a boolean representing whether or not a given tile blocks movement. We are also checking if the requested coordinates are smaller than 0 or larger than our Width and Height, and if so, to return as if blocking.
+Here we have defined our default Tile's flags. Note that we are using an [iota](https://golang.org/ref/spec#Iota) to both increment and bitshift the variables defined within the `const` block. What this means is that the first entry, **BlockMovement**, refers to the first bit in an integer, and the second entry of **BlockSight** refers to the second bit in an integer.
 
 ## Topics Covered
-We have created the `GameMap` type that uses our Tile type and will create few dummy tiles for testing.
+We have created a new package for holding types and functions corresponding to mapping and we have also defined a very basic Tile type.
 
 ## What's Next
-Now we will move on to integrating the `GameMap` with our main program.
+Now we can move to creating a `GameMap` type using our Tile type.
