+++ curriculum/part-2/7.go
@@ -1,65 +1,35 @@
-package main
+package mapping
 
-import (
-	"github.com/kettek/goro"
-	"log"
+// GameMap is our map data type.
+type GameMap struct {
+	Width, Height int
+	Tiles         [][]Tile
+}
 
-	"myproject/entity"
-	"myproject/mapping"
-)
+// Initialize initializes a GameMap's Tiles to match its Width and Height. It also sets up some coordinates to block movement and sight.
+func (g *GameMap) Initialize() {
+	g.Tiles = make([][]Tile, g.Width)
 
-func main() {
-	// Initialize goro!
-	if err := goro.InitTCell(); err != nil {
-		log.Fatal(err)
+	for x := range g.Tiles {
+		g.Tiles[x] = make([]Tile, g.Height)
 	}
 
-	goro.Run(func(screen *goro.Screen) {
-		// Screen configuration.
-		screen.SetTitle("My Roguelike")
+	g.Tiles[30][22] = Tile{
+		Flags: BlockMovement | BlockSight,
+	}
+	g.Tiles[31][22] = Tile{
+		Flags: BlockMovement | BlockSight,
+	}
+	g.Tiles[32][22] = Tile{
+		Flags: BlockMovement | BlockSight,
+	}
+}
 
-		// Our initial variables.
-		mapWidth, mapHeight := 80, 24
-
-		colors := map[string]goro.Color{
-			"darkWall":   goro.Color{R: 0, G: 0, B: 100, A: 255},
-			"darkGround": goro.Color{R: 50, G: 50, B: 150, A: 255},
-		}
-
-		gameMap := mapping.GameMap{
-			Width:  mapWidth,
-			Height: mapHeight,
-		}
-
-		gameMap.Initialize()
-
-		player := entity.NewEntity(screen.Columns/2, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorWhite})
-		npc := entity.NewEntity(screen.Columns/2-5, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorYellow})
-
-		entities := []*entity.Entity{
-			player,
-			npc,
-		}
-
-		for {
-			// Draw screen.
-			DrawAll(screen, entities, gameMap, colors)
-			ClearAll(screen, entities)
-
-			// Handle events.
-			switch event := screen.WaitEvent().(type) {
-			case goro.EventKey:
-				switch action := handleKeyEvent(event).(type) {
-				case ActionMove:
-					if !gameMap.IsBlocked(player.X+action.X, player.Y+action.Y) {
-						player.Move(action.X, action.Y)
-					}
-				case ActionQuit:
-					goro.Quit()
-				}
-			case goro.EventQuit:
-				return
-			}
-		}
-	})
+// IsBlocked returns if the given coordinates are blocking movement.
+func (g *GameMap) IsBlocked(x, y int) bool {
+	// Always block if outside our GameMap's bounds.
+	if x < 0 || x >= g.Width || y < 0 || y >= g.Height {
+		return true
+	}
+	return g.Tiles[y][x].Flags&BlockMovement != 0
 }
