+++ curriculum/part-2/8.go
@@ -1,44 +1,67 @@
 package main
 
 import (
+	"log"
+
 	"github.com/kettek/goro"
 
 	"myproject/entity"
+	"myproject/interfaces"
 	"myproject/mapping"
 )
 
-// DrawAll draws all entities and the gameMap to the screen and flushes it.
-func DrawAll(screen *goro.Screen, entities []*entity.Entity, gameMap mapping.GameMap, colors map[string]goro.Color) {
-	// Draw all the tiles within the game map.
-	for x, column := range gameMap.Tiles {
-		for y, tile := range column {
-			if tile.BlockSight {
-				screen.SetBackground(x, y, colors["darkWall"])
-			} else {
-				screen.SetBackground(x, y, colors["darkGround"])
+func main() {
+	// Initialize goro!
+	if err := goro.InitTCell(); err != nil {
+		log.Fatal(err)
+	}
+
+	goro.Run(func(screen *goro.Screen) {
+		// Screen configuration.
+		screen.SetTitle("My Roguelike")
+
+		// Our initial variables.
+		mapWidth, mapHeight := 80, 24
+
+		colors := map[string]goro.Color{
+			"darkWall":   goro.Color{R: 0, G: 0, B: 100, A: 255},
+			"darkGround": goro.Color{R: 50, G: 50, B: 150, A: 255},
+		}
+
+		gameMap := mapping.GameMap{
+			Width:  mapWidth,
+			Height: mapHeight,
+		}
+
+		gameMap.Initialize()
+
+		player := entity.NewEntity(screen.Columns/2, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorWhite})
+		npc := entity.NewEntity(screen.Columns/2-5, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorYellow})
+
+		entities := []interfaces.Entity{
+			player,
+			npc,
+		}
+
+		for {
+			// Draw screen.
+			DrawAll(screen, entities, gameMap, colors)
+			ClearAll(screen, entities)
+
+			// Handle events.
+			switch event := screen.WaitEvent().(type) {
+			case goro.EventKey:
+				switch action := handleKeyEvent(event).(type) {
+				case ActionMove:
+					if !gameMap.IsBlocked(player.X()+action.X, player.Y()+action.Y) {
+						player.Move(action.X, action.Y)
+					}
+				case ActionQuit:
+					goro.Quit()
+				}
+			case goro.EventQuit:
+				return
 			}
 		}
-	}
-	// Draw all the entities.
-	for _, entity := range entities {
-		DrawEntity(screen, entity)
-	}
-	screen.Flush()
-}
-
-// ClearAll clears all entities from the screen.
-func ClearAll(screen *goro.Screen, entities []*entity.Entity) {
-	for _, entity := range entities {
-		ClearEntity(screen, entity)
-	}
-}
-
-// DrawEntity draws a given entity to the screen.
-func DrawEntity(screen *goro.Screen, e *entity.Entity) {
-	screen.DrawRune(e.X, e.Y, e.Rune, e.Style)
-}
-
-// ClearEntity clears a given entity from the screen.
-func ClearEntity(screen *goro.Screen, e *entity.Entity) {
-	screen.DrawRune(e.X, e.Y, ' ', goro.Style{})
+	})
 }
