+++ curriculum/part-2/8.md
@@ -1,46 +1,65 @@
-# Rendering with the GameMap
-We must now return to our **render.go** file and make some changes so we can actually see our GameMap!
+# Using our GameMap
+Now we can move on to updating our main program to use our GameMap. Please note that we have yet to update our rendering to reflect the changes here, so rebuilding should be reserved for the end of the next lesson.
 
 ## Code Analysis
+
 ```
 import (
 	"github.com/kettek/goro"
+	"log"
 
 	"myproject/entity"
-+  "myproject/mapping"
+	"myproject/interfaces"
++	"myproject/mapping"
 )
 ```
-Here we have added our mapping package to our render file.
+Here we update our imports to include our new mapping package that GameMap resides within. Ensure that you replace **myproject** with whatever you named your module in [../part-1/1](Part 1:1).
 
 ---
 ```
--// DrawAll draws all entities to the screen and flushes it.
--func DrawAll(screen *goro.Screen, entities []*entity.Entity) {
-+// DrawAll draws all entities and the gameMap to the screen and flushes it.
-+func DrawAll(screen *goro.Screen, entities []*entity.Entity, gameMap mapping.GameMap, colors map[string]goro.Color) {
-+  // Draw all the tiles within the game map.
-+  for x, column := range gameMap.Tiles {
-+    for y, tile := range column {
-+      if tile.BlockSight {
-+        screen.SetBackground(x, y, colors["darkWall"])
-+      } else {
-+        screen.SetBackground(x, y, colors["darkGround"])
-+      }
-+    }
-+  }
-+	// Draw all the entities.
-	for _, entity := range entities {
-		DrawEntity(screen, entity)
-  }
+		// Our initial variables.
++		mapWidth, mapHeight := 80, 24
++
++		colors := map[string]goro.Color{
++			"darkWall":   goro.Color{R: 0, G: 0, B: 100, A: 255},
++			"darkGround": goro.Color{R: 50, G: 50, B: 150, A: 255},
++		}
++
++		gameMap := mapping.GameMap{
++			Width:  mapWidth,
++			Height: mapHeight,
++		}
++
++		gameMap.Initialize()
++
+		player := entity.NewEntity(screen.Columns/2, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorWhite})
+
 ```
-Here we have a fairly large addition that iterates through columns, then through each tile within the column. Within this iteration, we choose to either draw the "darkWall" or the "darkGround" background color if the tile in question blocks sight.
+Here we are adding quite a few things. The first is defining `mapWidth` and `mapHeight` variables. We are then defining a string [map](https://tour.golang.org/moretypes/19) to map color names to `goro.Color` types. After this, we create a `gameMap` variable with its Width and Height set to our earlier mapWidth and mapHeight variables. After this, we then call the gameMap's `Initialize()` method to ensure all data structures are set up properly.
 
-It is worth reminding here that golang's **for range** loop returns the index of the iteration followed by the actual type's value of the interation. This corresponds to `x, column :=` and `y, tile :=` as we are using both the indices and the data for both range loops.
+---
+```
+			// Draw screen.
+-			DrawAll(screen, entities)
++			DrawAll(screen, entities, gameMap, colors)
+			ClearAll(screen, entities)
+```
+Here we are updating our `DrawAll` function to reflect a change we have yet to make. This change passes both the gameMap and our map of colors we created earlier.
 
-With these changes in place, we can now rebuild the project and you should see much more color and be able to be blocked by the tiles rendered as walls.
+---
+```
+				case ActionMove:
+-					player.Move(action.X, action.Y)
++					if !gameMap.IsBlocked(player.X+action.X, player.Y+action.Y) {
++						player.Move(action.X, action.Y)
++					}
+```
+The final change we are making is to use the `GameMap.IsBlocked(...)` method we defined in the previous lesson. This will disallow the player from moving into an area that is considered to block movement.
+
+**NOTE**: If you build at this point, it will fail. The next lesson will bring our rendering up to date with our changes here.
 
 ## Topics Covered
-Here we have increased the functionality of our `DrawAll` function to allow rendering the game map.
+We have adjusted our main program to use our GameMap type as well as a new Color map.
 
 ## What's Next
-With this, we are done with part 2. Part 3 will consist of randomly generating a dungeon.
+Now we can update our rendering to match this file so we can actually render the GameMap and move our character around within it.
