+++ curriculum/part-2/9.go
@@ -4,10 +4,22 @@ import (
 	"github.com/kettek/goro"
 
 	"myproject/interfaces"
+	"myproject/mapping"
 )
 
-// DrawAll draws all entities to the screen and flushes it.
-func DrawAll(screen *goro.Screen, entities []interfaces.Entity) {
+// DrawAll draws all entities and the gameMap to the screen and flushes it.
+func DrawAll(screen *goro.Screen, entities []interfaces.Entity, gameMap mapping.GameMap, colors map[string]goro.Color) {
+	// Draw all the tiles within the game map.
+	for x, column := range gameMap.Tiles {
+		for y, tile := range column {
+			if tile.Flags&mapping.BlockSight != 0 {
+				screen.SetBackground(x, y, colors["darkWall"])
+			} else {
+				screen.SetBackground(x, y, colors["darkGround"])
+			}
+		}
+	}
+	// Draw all the entities.
 	for _, entity := range entities {
 		DrawEntity(screen, entity)
 	}
